refactor(fleet): split fleet modules under the 500-line cap
fleet.rs had reached 689 lines in the agent and 506 in the protocol, over the
repo's hard limit. Split along the seams that already existed rather than at
arbitrary line counts:

  protocol/fleet.rs   identity and link vocabulary
  protocol/fleet/node.rs   what a node reports about itself
  agent/fleet.rs      the fleet view and its state
  agent/fleet/listing.rs   turning that state into the list a browser renders
  agent/fleet/vitals.rs    where this machine's numbers come from

listing.rs carries the precedence rule in its module doc, because that is the
part a reader needs and it was previously buried: an authenticated peer report
beats a beacon sighting beats a remembered pin, and anything unpaired shows no
vitals at all.
