+++ crates/atlasctl-agent/src/fleet.rs
@@ -36,56 +36,6 @@ mod tests;
 /// missed refresh must not make a machine blink out of someone's interface.
 pub const UNREACHABLE_AFTER: Duration = Duration::from_secs(120);
 
-/// Supplies this machine's vitals.
-///
-/// A trait rather than a call into the telemetry functions directly, so the
-/// fleet view can be tested against a machine that answers everything, a
-/// machine that answers nothing, and the GB10 case where the memory questions
-/// have no answer at all.
-pub trait VitalsSource: Send + Sync {
-    /// The current sample.
-    fn vitals(&self) -> NodeVitals;
-}
-
-/// Turns a device sample into node vitals.
-///
-/// The `Option -> Metric` conversion is where "absent" is preserved: a field
-/// the hardware cannot answer becomes `Metric::Unsupported`, never zero.
-#[must_use]
-pub fn vitals_from_device(
-    d: &atlasctl_protocol::telemetry::DeviceStats,
-    disk_free_bytes: Option<f64>,
-    docker_ok: bool,
-    uptime_s: u64,
-    healthy_clock_mhz: Option<u32>,
-) -> NodeVitals {
-    use atlasctl_protocol::fleet::Metric;
-    let used_frac = match (d.memory_used_bytes, d.memory_total_bytes) {
-        #[expect(
-            clippy::cast_precision_loss,
-            reason = "a memory fraction is displayed to two significant figures"
-        )]
-        (Some(u), Some(t)) if t > 0 => Metric::reading(u as f64 / t as f64),
-        _ => Metric::Unsupported,
-    };
-    NodeVitals {
-        accelerator_util: d.gpu_util_pct.into(),
-        sm_clock_mhz: d.sm_clock_mhz.map(f64::from).into(),
-        sm_clock_healthy_mhz: healthy_clock_mhz,
-        temperature_c: d.temperature_c.into(),
-        power_w: d.power_w.into(),
-        memory_used_frac: used_frac,
-        #[expect(
-            clippy::cast_precision_loss,
-            reason = "byte counts are displayed in gigabytes"
-        )]
-        memory_total_bytes: d.memory_total_bytes.map(|b| b as f64).into(),
-        disk_free_bytes: disk_free_bytes.into(),
-        docker_ok,
-        agent_uptime_s: uptime_s,
-    }
-}
-
 /// What a session may ask about the fleet.
 pub trait FleetView: Send + Sync {
     /// Every node this agent knows about, local first.
@@ -115,9 +65,9 @@ pub struct PairOutcome {
 
 /// One entry in the discovery table.
 #[derive(Debug, Clone)]
-struct Seen {
-    beacon: Beacon,
-    at: Instant,
+pub(crate) struct Seen {
+    pub(crate) beacon: Beacon,
+    pub(crate) at: Instant,
 }
 
 /// This agent's view of the fleet.
@@ -305,7 +255,7 @@ impl LocalFleet {
         }
     }
 
-    fn lock_seen(&self) -> Option<MutexGuard<'_, BTreeMap<NodeId, Seen>>> {
+    pub(crate) fn lock_seen(&self) -> Option<MutexGuard<'_, BTreeMap<NodeId, Seen>>> {
         self.seen.lock().ok()
     }
 
@@ -346,177 +296,6 @@ impl LocalFleet {
             running: self.running.lock().ok().and_then(|r| r.clone()),
         }
     }
-}
-
-impl FleetView for LocalFleet {
-    fn nodes(&self) -> Vec<NodeDescriptor> {
-        let mut out = vec![self.local_node()];
-        let pinned = self.pins.load().unwrap_or_default();
-        let seen = self.lock_seen();
-        let alerts = self.alerts.lock().ok();
-
-        // Start from the pin store, so a paired machine that is switched off is
-        // still listed. A fleet that forgets its members when they sleep is not
-        // a fleet.
-        let reports = self.reports.lock().ok();
-        for (id, pin) in &pinned {
-            let sighting = seen.as_ref().and_then(|s| s.get(id));
-            let report = reports.as_ref().and_then(|r| r.get(id));
-            // A completed handshake is better evidence of liveness than a
-            // beacon in either direction: a wedged agent can still broadcast,
-            // and a healthy one goes quiet the moment a switch filters
-            // multicast.
-            let fresh =
-                report.is_some() || sighting.is_some_and(|s| s.at.elapsed() < UNREACHABLE_AFTER);
-            out.push(NodeDescriptor {
-                id: *id,
-                name: sighting.map_or_else(|| pin.name.clone(), |s| s.beacon.name.clone()),
-                is_local: false,
-                pairing: if fresh {
-                    PairingState::Paired
-                } else {
-                    PairingState::Unreachable
-                },
-                // A live sighting wins; otherwise fall back to where this peer
-                // was last known to be, so a restart does not make a paired
-                // machine look unreachable-and-addressless.
-                addresses: report
-                    .map(|r| {
-                        // Reached and authenticated, so the link class is ours
-                        // to state rather than a guess.
-                        vec![NodeAddress {
-                            iface: String::new(),
-                            addr: pin.last_address.clone().unwrap_or_default(),
-                            class: r.link,
-                            speed_mbps: None,
-                            rdma: matches!(
-                                r.link,
-                                atlasctl_protocol::fleet::LinkClass::Roce
-                                    | atlasctl_protocol::fleet::LinkClass::InfiniBand
-                            ),
-                        }]
-                    })
-                    .unwrap_or_else(|| {
-                        sighting.map_or_else(
-                            || {
-                                pin.last_address
-                                    .as_ref()
-                                    .map(|a| {
-                                        vec![NodeAddress {
-                                            iface: String::new(),
-                                            addr: a.clone(),
-                                            class: atlasctl_protocol::fleet::LinkClass::Unverified,
-                                            speed_mbps: None,
-                                            rdma: false,
-                                        }]
-                                    })
-                                    .unwrap_or_default()
-                            },
-                            |s| addresses_of(&s.beacon),
-                        )
-                    }),
-                launchability: sighting.map_or_else(
-                    || Launchability::no("not reachable right now"),
-                    |s| {
-                        if s.beacon.can_launch {
-                            Launchability::yes()
-                        } else {
-                            Launchability::no("this node reports it cannot run models")
-                        }
-                    },
-                ),
-                agent_version: String::new(),
-                accelerator: report.map_or_else(
-                    || sighting.map_or_else(String::new, |s| s.beacon.accelerator.clone()),
-                    |r| r.accelerator.clone(),
-                ),
-                // Only ever from the authenticated channel. A peer we have not
-                // spoken to reports none, rather than a beacon's word for it.
-                vitals: report.and_then(|r| r.vitals.clone()),
-                alerts: alerts
-                    .as_ref()
-                    .and_then(|a| a.get(id).cloned())
-                    .unwrap_or_default(),
-                running: None,
-            });
-        }
-
-        // Then anything seen but not pinned: visible, and trusted by nobody.
-        if let Some(seen) = seen.as_ref() {
-            for (id, s) in seen.iter() {
-                if pinned.contains_key(id) {
-                    continue;
-                }
-                out.push(NodeDescriptor {
-                    id: *id,
-                    name: s.beacon.name.clone(),
-                    is_local: false,
-                    pairing: PairingState::Discovered,
-                    addresses: addresses_of(&s.beacon),
-                    launchability: if s.beacon.can_launch {
-                        Launchability::yes()
-                    } else {
-                        Launchability::no("this node reports it cannot run models")
-                    },
-                    agent_version: String::new(),
-                    accelerator: s.beacon.accelerator.clone(),
-                    vitals: None,
-                    alerts: Vec::new(),
-                    running: None,
-                });
-            }
-        }
-        out
-    }
-
-    fn pair(&self, node: NodeId, code: &str) -> Result<PairOutcome> {
-        anyhow::ensure!(
-            crate::pairing::looks_like_code(code),
-            "a pairing code is {} digits",
-            crate::pairing::CODE_DIGITS
-        );
-        let seen = self
-            .lock_seen()
-            .and_then(|s| s.get(&node).cloned())
-            .ok_or_else(|| anyhow::anyhow!("that node is not visible on this network"))?;
-
-        // The ceremony itself runs over the peer channel; this is the point
-        // where that would be driven. Until the peer transport is wired in,
-        // refuse rather than pretend: a pairing that silently succeeded without
-        // a key exchange would write a pin that means nothing.
-        anyhow::bail!(
-            "cannot reach {} to pair: the peer channel is not connected",
-            seen.beacon.name
-        )
-    }
-
-    fn unpair(&self, node: NodeId) -> Result<bool> {
-        self.pins.remove(node)
-    }
-}
-
-/// Addresses a beacon advertised, as protocol addresses.
-///
-/// A beacon carries no link classification — it is unauthenticated, and letting
-/// a stranger tell you their link is RoCE would let them talk their way to the
-/// top of your preference order.
-///
-/// The class is therefore `Unverified` rather than a guess. Guessing "ethernet"
-/// was worse than saying nothing: it told the operator of a 200 Gb RoCE fabric
-/// that their cluster would run on ethernet. It is resolved once the peer is
-/// reached over the authenticated channel and describes itself.
-fn addresses_of(beacon: &Beacon) -> Vec<NodeAddress> {
-    beacon
-        .addresses
-        .iter()
-        .map(|a| NodeAddress {
-            iface: String::new(),
-            addr: a.to_string(),
-            class: atlasctl_protocol::fleet::LinkClass::Unverified,
-            speed_mbps: None,
-            rdma: false,
-        })
-        .collect()
 }
 
 /// Record a completed pairing.
@@ -568,122 +347,7 @@ pub fn no_vitals() -> NodeVitals {
     NodeVitals::default()
 }
 
-/// Vitals read from this machine.
-///
-/// Capabilities are probed once at construction, because the answer does not
-/// change while the agent runs and re-probing every second would spawn a
-/// process every second. Individual readings are taken per sample.
-pub struct SystemVitals {
-    runner: std::sync::Arc<dyn atlasctl_core::io::ProcessRunner>,
-    caps: atlasctl_protocol::telemetry::TelemetryCaps,
-    started: Instant,
-    /// Filesystem whose free space matters — images and the model cache.
-    disk_path: std::path::PathBuf,
-}
+mod listing;
+pub mod vitals;
 
-impl std::fmt::Debug for SystemVitals {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        f.debug_struct("SystemVitals")
-            .field("caps", &self.caps)
-            .finish_non_exhaustive()
-    }
-}
-
-impl SystemVitals {
-    /// Probe this machine's telemetry capabilities and start sampling.
-    #[must_use]
-    pub fn new(
-        runner: std::sync::Arc<dyn atlasctl_core::io::ProcessRunner>,
-        disk_path: std::path::PathBuf,
-    ) -> Self {
-        let caps = crate::telemetry::probe(runner.as_ref());
-        Self {
-            runner,
-            caps,
-            started: Instant::now(),
-            disk_path,
-        }
-    }
-
-    /// What this machine can answer.
-    #[must_use]
-    pub const fn caps(&self) -> &atlasctl_protocol::telemetry::TelemetryCaps {
-        &self.caps
-    }
-}
-
-impl VitalsSource for SystemVitals {
-    fn vitals(&self) -> NodeVitals {
-        // `busy` drives the clamp decision, and it is the agent's call rather
-        // than the client's: an idle part at a low clock is normal, the same
-        // clock under load is the failure that hides for weeks.
-        let device = crate::telemetry::sample_device(self.runner.as_ref(), &self.caps, false);
-        let disk = free_bytes(&self.disk_path);
-        let docker_ok = self
-            .runner
-            .run(&docker_probe_argv())
-            .is_ok_and(|o| o.success());
-        vitals_from_device(
-            &device,
-            disk,
-            docker_ok,
-            self.started.elapsed().as_secs(),
-            self.caps.sm_clock_healthy_mhz,
-        )
-    }
-}
-
-/// The one way this project asks whether the container runtime is answering.
-///
-/// `docker info` exposes `.ServerVersion`; `docker version` does not — its
-/// field is `.Server.Version`, and asking `version` for `.ServerVersion` exits
-/// non-zero with a template error on Docker 29. That is a silent way to report
-/// a healthy daemon as unreachable, so there is exactly one definition of the
-/// probe and both callers use it.
-#[must_use]
-pub fn docker_probe_argv() -> Vec<String> {
-    vec![
-        "docker".to_owned(),
-        "info".to_owned(),
-        "--format".to_owned(),
-        "{{.ServerVersion}}".to_owned(),
-    ]
-}
-
-/// Free bytes on the filesystem holding `path`, when it can be determined.
-///
-/// A full model cache is a leading cause of launch failure, so this is worth
-/// reporting even though it needs a platform call.
-fn free_bytes(path: &std::path::Path) -> Option<f64> {
-    #[cfg(unix)]
-    {
-        use std::os::unix::ffi::OsStrExt;
-        // The cache directory may not exist yet on a fresh install, and statvfs
-        // on a missing path fails. What matters is the filesystem that WOULD
-        // hold it, so walk up to the nearest ancestor that does exist.
-        let mut probe = path;
-        while !probe.exists() {
-            match probe.parent() {
-                Some(parent) => probe = parent,
-                None => return None,
-            }
-        }
-        let c = std::ffi::CString::new(probe.as_os_str().as_bytes()).ok()?;
-        // SAFETY: `c` is a valid NUL-terminated path and `stat` is written only
-        // by the call, which reports success before we read it.
-        let mut stat: libc::statvfs = unsafe { std::mem::zeroed() };
-        if unsafe { libc::statvfs(c.as_ptr(), &raw mut stat) } != 0 {
-            return None;
-        }
-        #[expect(
-            clippy::cast_precision_loss,
-            reason = "free space is displayed in gigabytes"
-        )]
-        Some(stat.f_bavail as f64 * stat.f_frsize as f64)
-    }
-    #[cfg(not(unix))]
-    {
-        let _ = path;
-        None
-    }
-}
+pub use vitals::{SystemVitals, VitalsSource, docker_probe_argv, vitals_from_device};
